Balance Windows test parallelism
